change compare and print; optional type safety
having compare and print operate on a node of the tree was a dumb idea,
and makes generalizing the tree to other data structures difficult.
Have those routines instead take the key, which is what we care about
after all.

In the process, I got tripped up by all the void *, and added an option
to instantiate the tree for a particular type, which you can do by
defining the RBTREE_TYPE to something before including rbtree.h