fix payload copy in delete case
CLRS algorithim for deletion will in some cases pick a sucessor node and
copy data from node to be deleted into the successor in the tree and
point to a different node to delete.  I omitted that case in general and
definitely for the augmented codes.   Plus, lots more comments so I can
remember what I re-re-learned